Moved DVAddr out of zfs.rs
